+++ RestroHub/src/main/java/com/restroly/qrmenu/payment/service/PaymentService.java
@@ -1,14 +1,16 @@
 package com.restroly.qrmenu.payment.service;
 
+import java.math.BigDecimal;
+
 import org.springframework.core.io.Resource;
 
 import jakarta.transaction.Transactional;
 
 public interface PaymentService {
     @Transactional
-    String newPayment(Long orderId, double amount);
-    String generatePaymentLink(double amount, Long orderId, String upiId);
-    Resource generateUPIQR(double amount, String upiId, String description);
+    String newPayment(Long orderId, BigDecimal amount);
+    String generatePaymentLink(BigDecimal amount, Long orderId, String upiId);
+    Resource generateUPIQR(BigDecimal amount, String upiId, String description);
     @Transactional
     void markPaymentAsVerified(String paymentId, String transactionId);
     @Transactional
